TEMPORARY: cmake: Force to link all llvm modules
This is a quick hack for cross-compiling bcc,
and needs to be reworked.

When cross-compiling, cmakefiles in current seems to
try to dropping several llvm modules which determined
to be unnecessary for the target architecture.

This hack forces to link all llvm modules which have
been installed to the target system.